requirements.txt
amqp==2.5.0
bs4==0.0.1
celery==4.3.0
coverage==4.5.3
Django==2.2.3
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
flake8==3.7.7
gunicorn==19.9.0
Markdown==3.1.1
mysqlclient==1.4.2.post1
psycopg2==2.8.3
pycodestyle==2.5.0
PyMySQL==0.9.3
requests==2.22.0
requests-mock==1.6.0
